requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
discord.py==2.0.1
exceptiongroup==1.0.0rc9
frozenlist==1.3.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==4.9.1
multidict==4.7.6
outcome==1.2.0
Pillow==9.3.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
rfc3986==1.5.0
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
wsproto==1.2.0
yarl==1.5.1

discord~=2.1.0
